fix readme, add info on networking configuration
Since there are hundreds of Project Euler problems, and solution
environments may be similar -- it makes no sense to design ip
addressing for environments to run at the same time.

But there is a compelling reason to allow someone to start from
any solution, and easily get up and running.

Yes, vagrant may be "heavy", but it is simple and  proven.
Containers could be simpler, but the technology is evolving.